Support multitargeting (#4)
Re-work the targets to support when the referencing project and / or the
leaf project are multitargeting.

The process is now as follows:

1. Collect the references to pack and ensure they have
`GeneratePackageOnBuild`.
2. If the main project is multi-targeting, call Build (which calls Pack)
3. If the main project is single-targeting, call Build (which calls
Pack)
4. Query the output items and create item metadata

GeneratePackageOnBuild can't be used when the leaf is multitargeting
since Pack is an outer target and thus inner builds don't trigger a
Pack. However, this property is enforced as part of an opt-in / contract
in the referenced project to ensure it is designed to be packed as part
of the build.