Also call Pack in GetPackFromProject_PackFor* (#10)
The GetPackFromProject_PackFor* targets were calling `Build` on the leaf
project and assuming that both `Build` and `Pack` would be done (relying
on `GeneratePackageOnBuild` to do the heavy lifting). It appears that
was never a safe assumption to make, but the race seemed to reliably
work out that way.

With .NET 9 SDKs, it appears that we're now sometimes losing the race.
The leaf project builds, returning control to PackFor*, and then the
referencing project builds assuming the .nupkg is already available.
However, the referencing build completes before the leaf Pack completes,
and thus the copy to output directory fails, resulting in an error like
this:

> Error Could not copy the file
"D:\a\moq.analyzers\moq.analyzers\artifacts\package\release\Moq.Analyzers.0.3.0-alpha-ge454d76cff.nupkg"
because it was not found.

The fix is to explicitly call Pack as well to ensure its complete before
continuing.